requirements.txt
Django==1.11
django-extensions==1.7.8
django-filter==1.0.2
django-webpack-loader==0.4.1
djangorestframework==3.6.2
gunicorn==19.7.1
Markdown==2.6.8
whitenoise